слегонца подправил интерфейс
