--- settings.properties
@@ -1,5 +1,0 @@
-#terminal settings
-#Wed Mar 10 18:45:21 MSK 2021
-Id=80
-BoundRate=19200
-ComPort=COM4
