+++ src/ru/sid/izk/modbus/frames/IZKModbusGUI.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="IZKModbusGUI">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="ru.sid.izk.modbus.frames.IZKModbusGUI">
   <grid id="27dc6" binding="mainPanel" layout-manager="GridLayoutManager" row-count="13" column-count="6" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -365,17 +365,6 @@
             <properties/>
             <border type="none"/>
             <children>
-              <component id="b7627" class="javax.swing.JButton" binding="reButton2">
-                <constraints>
-                  <grid row="0" column="0" row-span="1" col-span="2" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="0" use-parent-layout="false">
-                    <preferred-size width="150" height="30"/>
-                    <maximum-size width="150" height="30"/>
-                  </grid>
-                </constraints>
-                <properties>
-                  <text value="Обновить данные"/>
-                </properties>
-              </component>
               <component id="64851" class="javax.swing.JLabel">
                 <constraints>
                   <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="4" fill="0" indent="0" use-parent-layout="false"/>
