+++ src/ru/sid/izk/modbus/frames/IZKTerminal.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="IZKTerminal">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="ru.sid.izk.modbus.frames.IZKTerminal">
   <grid id="27dc6" binding="terminalPanel" layout-manager="GridLayoutManager" row-count="5" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="5"/>
     <constraints>
