+++ src/ru/sid/izk/modbus/utils/DigitFilter.java
@@ -1,10 +1,13 @@
+package ru.sid.izk.modbus.utils;
+
 import javax.swing.text.AttributeSet;
 import javax.swing.text.BadLocationException;
 import javax.swing.text.DocumentFilter;
 
 public class DigitFilter extends DocumentFilter {
-    private int count;
-private static final String DIGITS = "\\d+";
+
+    private final int count;
+    private static final String DIGITS = "\\d+";
 
     public DigitFilter(int count) {
         this.count = count;
@@ -12,11 +15,11 @@ private static final String DIGITS = "\\d+";
 
     @Override
     public void insertString(FilterBypass fb, int offset, String string, AttributeSet attr) throws BadLocationException {
-        if (string.matches(DIGITS)) super.insertString(fb,offset,string,attr);
+        if (string.matches(DIGITS)) super.insertString(fb, offset, string, attr);
     }
 
     @Override
     public void replace(FilterBypass fb, int offset, int length, String text, AttributeSet attrs) throws BadLocationException {
-        if (text.matches(DIGITS) && offset<count) super.replace(fb, offset, length, text, attrs);
+        if (text.matches(DIGITS) && offset < count) super.replace(fb, offset, length, text, attrs);
     }
 }
