+++ src/ru/sid/izk/modbus/utils/FloatDigitalFilter.java
@@ -1,3 +1,5 @@
+package ru.sid.izk.modbus.utils;
+
 import javax.swing.text.AttributeSet;
 import javax.swing.text.BadLocationException;
 import javax.swing.text.DocumentFilter;
